The button doesn't always receive the impulse, but i don't know why, i will ask around

+++ src/changeStateTask.h
@@ -19,7 +19,7 @@ class ChangeStateTask {
 
     private:
     enum {MAN, AUTO} state;
-    int currentAngle;
+    double currentAngle;
 
 };
